Convert package to quilt format

+++ caPutLogApp/caPutLogTask.c
@@ -179,7 +179,7 @@ void caPutLogTaskSend(LOGDATA *plogData)
 static void caPutLogTask(void *arg)
 {
     int sent = FALSE, burst = FALSE;
-    int config = (int)(size_t)arg;
+    int config = (size_t)arg;
     LOGDATA *pcurrent, *pnext;
     VALUE old_value, max_value, min_value;
     VALUE *pold=&old_value, *pmax=&max_value, *pmin=&min_value;
